Finish

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: bahkaya <bahkaya@student.42istanbul.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/12 16:30:54 by bahkaya           #+#    #+#             */
-/*   Updated: 2025/06/15 15:59:10 by bahkaya          ###   ########.fr       */
+/*   Updated: 2025/06/19 16:48:34 by bahkaya          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*str;
 
 	i = 0;
